refactor: Rename `._reportSdkError` to `._reportError`
As this utility is used to report both SDK and user errors

BREAKING CHANGE:
`serverless._reportSdkError` is renamed to `serverless._reportError`
