+++ node/packages/sdk/lib/report-error.js
